requirements.txt
azure-iot-device
pymodbus
ocpp
pytest
pytest-asyncio